Reworked schema, added "override-codec" sigs, fixed example
Java "object/class types" are no longer allowed. Use generic "O" instead.

"related-methods" was added: "A list of ASM signatures that are in the same class to help the patcher find the method.
This is optional, but this may help if a method has very few opcodes."

Also updated "$id" in schema to use the new org name.

The old "signatures.json" file was moved to "signatures.base.json".